relocate: validate_against function to validation.py
