+++ src/python/impactx/dashboard/Input/__init__.py
@@ -3,6 +3,7 @@ from .components.input import InputComponents
 from .components.navigation import NavigationComponents
 from .defaults import DashboardDefaults
 from .generalFunctions import generalFunctions
+from .validation import DashboardValidation
 
 __all__ = [
     "CardBase",
@@ -10,6 +11,7 @@ __all__ = [
     "InputComponents",
     "NavigationComponents",
     "DashboardDefaults",
+    "DashboardValidation",
     "generalFunctions",
     "InputComponents",
     "NavigationComponents",
