+++ src/python/impactx/dashboard/Input/generalFunctions.py
@@ -77,58 +77,6 @@ class generalFunctions:
                 return None
 
     @staticmethod
-    def validate_against(input_value, value_type, additional_conditions=None):
-        """
-        Validates the input value against the desired type and additional conditions.
-        :param input_value: The value to validate.
-        :param value_type: The desired type ('int', 'float', 'str').
-        :param additional_conditions: A list of additional conditions to validate.
-        :return: A list of error messages. An empty list if there are no errors.
-        """
-        errors = []
-        value = None
-
-        if input_value == "None":
-            return errors
-
-        # value_type checking
-        if value_type == "int":
-            if input_value is None:
-                errors.append("Must be an integer")
-            else:
-                try:
-                    value = int(input_value)
-                except ValueError:
-                    errors.append("Must be an integer")
-        elif value_type == "float":
-            if input_value is None:
-                errors.append("Must be a float")
-            else:
-                try:
-                    value = float(input_value)
-                except ValueError:
-                    errors.append("Must be a float")
-        elif value_type == "str":
-            if input_value is None:
-                errors.append("Must be a string")
-            else:
-                value = str(input_value)
-        else:
-            errors.append("Unknown type")
-
-        # addition_conditions checking
-        if errors == [] and additional_conditions:
-            for condition in additional_conditions:
-                if condition == "non_zero" and value == 0:
-                    errors.append("Must be non-zero.")
-                if condition == "positive" and value < 0:
-                    errors.append("Must be positive.")
-                if condition == "negative" and value > 0:
-                    errors.append("Must be negative.")
-
-        return errors
-
-    @staticmethod
     def update_simulation_validation_status():
         """
         Checks if any input fields are not provided with the correct input type.
