+++ src/python/impactx/dashboard/Input/shared.py
@@ -8,7 +8,7 @@ License: BSD-3-Clause-LBNL
 
 from .. import setup_server
 from ..Input.inputParameters.inputMain import InputParameters
-from . import DashboardDefaults, generalFunctions
+from . import DashboardDefaults, DashboardValidation, generalFunctions
 
 server, state, ctrl = setup_server()
 
@@ -32,7 +32,7 @@ class SharedUtilities:
                     state_name, None
                 )
 
-                validation_result = generalFunctions.validate_against(
+                validation_result = DashboardValidation.validate_against(
                     input, desired_type, conditions
                 )
                 setattr(state, validation_name, validation_result)
